% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merge datasets}
\usage{
merge_datasets(directory = getwd(), file_ext = "tab", step = "between")
}
\arguments{
\item{directory=getwd()}{Working directory with delimited files.}

\item{file_ext="tab"}{File extension of delimited files containing
M, name and SPOT columns.}

\item{step="between"}{Determine step of normalization: "between" means
between array step.}
}
\value{
A tibble containing M, name and SPOT columns, if step = "between",
or CH1, CH2, name and SPOT in other cases.
}
\description{
Merge *.tab datasets from a directory. Safe them as tibble with
M, name and SPOT columns. Secure switching back to the user-defined
file directory.
}
