% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(microarray)
}
\arguments{
\item{mocroarray}{A table containing at least M, A and name columns.}
}
\value{
Same table normalized and with NA removed.
}
\description{
Normalize mean and average.
}
