% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{prepare_single_microarray}
\alias{prepare_single_microarray}
\title{File preparation}
\usage{
prepare_single_microarray(file_name)
}
\arguments{
\item{file_name}{Name of a delimited file containing
at least SPOT, CH1B, CH1I, CH2I and
CH2B columns.}
}
\value{
Tibble containing SPOT, CH1B, CH1I, CH2I,
CH2B and name columns.
}
\description{
Prepare a single microarray dataset for plotting.
Reading data and selecting columns of interest.
}
