% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{purge}
\alias{purge}
\title{Purge}
\usage{
purge(microarray)
}
\arguments{
\item{microarray}{A tibble containing M, name and SPOT columns.}
}
\value{
A tibble containing only  M, name and SPOT columns.
}
\description{
Remove all columns except M, name and SPOT.
}
