% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run App}
\usage{
run_app(gui = ui, gserver = server)
}
\arguments{
\item{gui}{ui argument of a shiny app.}

\item{gserver}{server argument of a shiny app.}
}
\description{
Run Shiny App for more comfortable plotting.
}
