% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{scale_norm}
\alias{scale_norm}
\title{Scaling}
\usage{
scale_norm(microarrays)
}
\arguments{
\item{microarrays}{A tibble containing M and name columns.}
}
\value{
Same tibble with  M column scaled.
}
\description{
Scaling step of between array normalization.
Divide each value through MAD, so std. deviations become 1.
}
