% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{subtract_background}
\alias{subtract_background}
\title{Remove noise}
\usage{
subtract_background(microarray)
}
\arguments{
\item{microarray}{A tibble containing at least SPOT, CH1B,
CH1I, CH2I, CH2B and name columns.}
}
\value{
Tibble containing input columns and two
new columns CH1 and CH2. All values <= 0
become 1.
}
\description{
Subtract background noise from every channel.
}
